user and admin changes
